Resetting the game flow.
Now made it so that when you exit the game and enter back in you restart. Same as if you lose all your lives or the police car collides with the player car which ends the game which takes you to the game over screen which shows you your score this time and your high score, that game over screen is blank currently, when you go back in it starts you back at the start instead of leaving you where you left off.