Working on GameOverScreen and startscreen
I've got the score showing and the highscore on the gameoverscreen, plus the high score is being displayed on the startscreen. But the highscore being displayed is the top speed of the car. need to fix that and position things on the game over screen. Also gonna make it so that the kilometers travelled shows up at the top of playscreen again.